put moveEvidence into evidence file
